Several improvements
I've made a new class to handle the input data, so it's now able to process notations inpependent upon the list structure. Also, it is now able to process the content of [] brackets as metadata containers. Now, it's able to read the names of players and display them on screen.