Preventing abuse of regexps and exec statements
  In this commit I have created the class `Move` inside managers.py, and
implemented it to replace the messy list instances that stored the parameters
for every move.

  Also, thanks to this, I could finally stop the stupid abuse on regexps and
remove the `exec` statements in `funciones.py`.

  This is a huge step towards a better modelling of the software.

  Once more I want to thank @200_success and @jonrsharpe, who helped me
through codereview.com to see the modelling lacks of the software's whole
structure and gave me awesome pointers to improve the source code. I haven't
implemented all their suggestions yet, but after this commit I will be able to
do it more easily in the future. http://codereview.stackexchange.com/questions/96707/class-modelling-for-a-shogi-notation-reader

  On the next commit I will harness this new data structure to create loops
that will eliminate ridiculously redundant code.